Comprehensive build fix: remove SDK pin, fix include order, overhaul CI workflow
- CMakeLists.txt: Remove CMAKE_SYSTEM_VERSION pin (caused SDK mismatch on CI),
  add DirectX-Headers include dir ordering (BEFORE system SDK) to prevent
  include guard conflicts with vcpkg headers
- build.yml: Pin windows-2022, add ilammy/msvc-dev-cmd, explicit VS generator,
  parallel builds, improved caching, conditional asset copy
- ui_context.h: Forward declarations only (no DX12 header leaks)
- ui_context.cpp: Heavy DX12/ImGui includes moved here from header
- pipeline_view.h, calibration_panel.h: Direct imgui.h includes
- main.cpp: Explicit imgui_impl_dx12.h and imgui_impl_win32.h includes
- vcpkg.json: Add builtin-baseline for reproducible builds

+++ RobotSimulator_CPP/ui/ui_context.cpp
@@ -1,8 +1,9 @@
 #include "ui_context.h"
+
+// Heavy includes — only in the .cpp, NOT in header
+#include "../engine/dx12_core.h"
 #include <imgui_impl_win32.h>
 #include <imgui_impl_dx12.h>
-#include "imgui.h"
-#include "../engine/dx12_core.h"
 #include "pipeline_view.h"
 #include "calibration_panel.h"
 
@@ -48,7 +49,6 @@ void UIContext::Render() {
     RenderDiagnosticsPanel();
     if (m_pipelineView) m_pipelineView->Render();
     if (m_calibrationPanel) m_calibrationPanel->Render();
-    // RenderMenu();
 }
 
 void UIContext::EndFrame() {
@@ -107,7 +107,7 @@ void UIContext::RenderDiagnosticsPanel() {
                 ImGui::TableNextRow();
                 
                 ImGui::TableNextColumn();
-                ImGui::Text("J%d", i + 1);
+                ImGui::Text("J%d", (int)(i + 1));
                 
                 ImGui::TableNextColumn();
                 ImGui::Text("%.4f", ref[i]);
